requirements.txt
google-genai
pandas
openpyxl
streamlit
numpy
